refactor(util): 패키지 분리
