--- src/test/java/ladder/domain/RandomBooleanGeneratorTest.java
@@ -1,7 +1,0 @@
-package ladder.domain;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class RandomBooleanGeneratorTest {
-
-}
