+++ src/test/java/ladder/util/BooleanGeneratorTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.util;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
